+++ src/DotNetStac/Extensions/Eo/EoBandCommonName.cs
@@ -5,7 +5,6 @@ namespace Stac.Extensions.Eo
     /// </summary>
     public enum EoBandCommonName
     {
-        unknow,
         coastal,
         blue,
         green, 
@@ -13,28 +12,14 @@ namespace Stac.Extensions.Eo
         yellow,
         pan,
         rededge,
-        rededge70,
-        rededge74,
-        rededge76,
-        rededge78,
         nir,
         nir08,
         nir09,
-        swir12,
         cirrus, 
         swir16,
-        swir155,
-        swir165,
-        swir173,
         swir22,
-        swir215,
-        swir220,
-        swir225,
-        swir23,
-        swir37,
         lwir,
         lwir11,
-        lwir12,
-        alpha
+        lwir12
     }
 }